Rainting on reviews and price tier on all master cells and screens

+++ Beautify/Model/Review.swift
@@ -15,5 +15,5 @@ struct Review {
     var masterID: String?
     var topImageURL: String?
     var description: String?
-    var grade: Int?
+    var grade: Double?
 }
