+++ Beautify/ViewModel/Master/Card/MasterCardViewModel.swift
@@ -105,6 +105,6 @@ struct ReviewViewModel {
     var masterID: String?
     var topImageURL: String?
     var description: String?
-    var grade: Int?
+    var grade: Double?
 }
